Move hook functions to hookTestUtils
